move cache helpers to utils
