+++ backend/graphrag_app/api/data.py
@@ -5,10 +5,8 @@ import asyncio
 import csv
 import re
 import traceback
-from hashlib import sha256
-from io import StringIO
 from math import ceil
-from typing import BinaryIO, List
+from typing import List
 
 from azure.storage.blob.aio import ContainerClient
 from fastapi import (
@@ -25,11 +23,13 @@ from graphrag_app.typing.models import (
     StorageNameList,
 )
 from graphrag_app.utils.common import (
+    check_cache,
     delete_cosmos_container_item_if_exist,
     delete_storage_container_if_exist,
     get_blob_container_client,
     get_cosmos_container_store_client,
     sanitize_name,
+    update_cache,
 )
 
 data_route = APIRouter(
@@ -90,45 +90,6 @@ async def upload_file_async(
                 await update_cache(filename, file_stream, container_client)
         except Exception:
             pass
-
-
-async def check_cache(file_stream: BinaryIO, container_client: ContainerClient) -> bool:
-    """
-    Check if the file has already been uploaded.
-    """
-    # load the file cache
-    cache_blob_client = container_client.get_blob_client("uploaded_files_cache.csv")
-    cache_download_stream = await cache_blob_client.download_blob()
-    cache_bytes = await cache_download_stream.readall()
-    cache_content = StringIO(cache_bytes.decode("utf-8"))
-
-    # comupte the sha256 hash of the file and check if it exists in the cache
-    cache_reader = csv.reader(cache_content, delimiter=",")
-    file_hash = sha256(file_stream.read()).hexdigest()
-    for row in cache_reader:
-        if file_hash in row:
-            return True
-    return False
-
-
-async def update_cache(filename: str, file_stream: BinaryIO, container_client: ContainerClient) -> None:
-    """
-    Update the file cache with the new file.
-    """
-    # load the file cache
-    cache_blob_client = container_client.get_blob_client("uploaded_files_cache.csv")
-    cache_download_stream = await cache_blob_client.download_blob()
-    cache_bytes = await cache_download_stream.readall()
-    cache_content = StringIO(cache_bytes.decode("utf-8"))
-
-    # compute the sha256 hash of the file and add it to the cache
-    cache_writer = csv.writer(cache_content, delimiter=",")
-    file_hash = sha256(file_stream.read()).hexdigest()
-    cache_writer.writerow([filename, file_hash])
-
-    # upload the updated cache to Azure Blob Storage
-    cache_content.seek(0)
-    await cache_blob_client.upload_blob(cache_content, overwrite=True)
 
 
 class Cleaner:
